move some runner functions to shared file
